Add support for parameter grouping
Introduces the BuildParameterGroup concept which can be used to group
related build parameters. This results in the groups parameters being
accessible under a common group prefix both when accessed in a build
script and when overridden on the command line.
As part of this the BuildParametersExtension now inherits from
BuildParameterGroup and there is some special handling to detect this
when a group or parameter is added.