Check for supported Gradle version (#50)
The plugin now checks whether it's running on a supported Gradle version.
Only versions from Gradle 7.1 on wards are supported because previous versions
required calling `forUseAtConfigurationTime()` on providers that are resvoled
at configuration time. Later on that method was deprecated again. While it would
be possible to support older versions, it would lead to quite some complexity
in the generated plugin's code.

Currently only the generator plugin contains a version check, but not the
generated plugin. Since it's theoretically possible to publish the generated
plugin to a repository and then use it on an older Gradle version, there's
some potential for late runtime errors. So adding the check to the generated
plugin is a potential follow up improvement.

Last but not least this adds a cross version test as part of the default
test suite that verifies the plugin works on the minimum supported version
and the highest patch version for each Gradle minor release.

Resolves #17, #18